# Speed up bound checking in `take`