--- a/arrow/src/compute/kernels/take.rs
+++ b/arrow/src/compute/kernels/take.rs
@@ -17,7 +17,7 @@
 
 //! Defines take kernel for [Array]
 
-use std::{ops::AddAssign, sync::Arc, usize};
+use std::{ops::AddAssign, sync::Arc};
 
 use crate::buffer::{Buffer, MutableBuffer};
 use crate::compute::util::{
@@ -276,19 +276,6 @@ fn maybe_usize<I: ArrowPrimitiveType>(index: I::Native) -> Result<usize> {
         .ok_or_else(|| ArrowError::ComputeError("Cast to usize failed".to_string()))
 }
 
-fn take_no_nulls_u32<T>(
-    values: &[T::Native],
-    indices: &[u32],
-) -> Result<(Buffer, Option<Buffer>)>
-where
-    T: ArrowPrimitiveType,
-{
-    let values = indices.iter().map(|index| values[*index as usize]);
-    let buffer = unsafe { Buffer::from_trusted_len_iter(values) };
-
-    Ok((buffer, None))
-}
-
 // take implementation when neither values nor indices contain nulls
 fn take_no_nulls<T, I>(
     values: &[T::Native],
@@ -446,15 +433,7 @@ where
         (false, false) => {
             // * no nulls
             // * all `indices.values()` are valid
-            if I::DATA_TYPE == DataType::UInt32 {
-                let indices_u32 = indices
-                    .as_any()
-                    .downcast_ref::<PrimitiveArray<UInt32Type>>()
-                    .expect("");
-                take_no_nulls_u32::<T>(values.values(), indices_u32.values())?
-            } else {
-                take_no_nulls::<T, I>(values.values(), indices.values())?
-            }
+            take_no_nulls::<T, I>(values.values(), indices.values())?
         }
         (true, false) => {
             // * nulls come from `values` alone